Fix incorrect alignment of new id
The create_id func separately created a new building id for different
files (e.g. _attrib.csv or _geom.csv) while relying on the ordering of
buildings in the files to ensure consistency between them.

This resulted in inconsistencies as the ordering differed between the
files and the mapping of the old id_source and new id column was broken
across files.

Now, the new id is solely created for the _attrib.csv and other files
are aligned based on the mapping of id_source and id.

(cherry picked from commit 04204b4d0e4b009518ba9bf171014bac7cec3ce3)