Adds folders when creating unique ids and tests for id errors.
When creating new ids, folders will be created when they do not exist yet. This is the case
f.e. when creating a new release version. In addition, we add tests to check that for every city
we have the same number of entries across all files (geom, attrib, extra_attrib, source_file) and
no id or id_source duplicates in any of the files.