Creates folders only for cities of dataset_name.
This addresses the problem of not being able to create folders
if they already exist as previously folders for whole country were
created. Now only the folders (f.e. all folders of Wien for wien-gov
dataset) are created and added to the paths.txt.

This also affects the following:
- boundary_geoms are also only created in cities to the respective
dataset_name
- masking of gadm happens at the beginning and is not done anymore
when creating building geometries.

Still open:
- we should think about if overwrite = True, should we first delete
all files in the respective folders?