Update parsing.py
Introduced the id_counter into extra_attrib where it had been previously forgotten. This was causing future merges between attrib and extra attrib to fail as some buildings had different ids.

This counter enables to have a continuous index for a dataset when there are multiple chunks, where id: <source_file>_<index>. (Not sure why this is here, but presumably for cases with multiple source files having the same name, not to cause id duplicates.)