Fix dropping of duplicates for intersecting buildings
For example, a OSM building can intersect with multiple gov buildings;
to determine if the overlap is sufficiently small to add the OSM
building to the main dataset, we are only interested in the overlap with
the gov building with the largest intersection. All other intersection
pairs have to be dropped.