Adds case Bremen and Berlin.
Bremen requires several additional geometry steps as gml files either have wkb errors or geometry errors.
In addition, geometries are provided as multiline strings with a Z geometry.

The gml files are now read in first with engine = fiona and if that fails we try
engine=pygrio. If both fail the filename is reported in the log file with the schema
"GML ERROR. Could not read: {file}" and the next file is read in.

In addition, any multiline string geometry is converted to Polygon or Multipolygon and any z geometry is reduced to a 2d geometry.

Future versions of this should additional check if a height columen is present in the data and if not take thez value of the polygon as height.