Improves reading in of gml files.
For some geometries in case Bremen, reading a gml file does throw an error.
This is due to the fact that while fiona reads in the file, a meta file in .gfs format
is created that can help to read it in. However, in some cases the creation of .gfs
takes too long and gpd.read_file does not consider it. That is why we follow the logic:
1. try reading in using engine=fiona
2. if it breaks, try reading in using pygrio
3. if it breaks, wait 0.5 sec (to ensure .gfs creation) and try reading in again with fiona
4. if it breaks, report error and move to next file