add testing case for bavaria which handles gml files differently.
The cluster struggles to directly read a url via geopandas.read_file().
While this works without issues on macOS it fails on the cluster.
Potential solutions with different suopported drivers in fiona seem not to
solve the issue as they are 100% the same. Similarly all packages are the same
version and are installed the same way on both machines.