Read in safely None values from input parsing.
When running parsing on the cluster, the following error occured:
"ValueError: malformed node or string: np.float64(nan)"
This is likely due to the fact that newer pandas versions are inter-
preting None not as string anymore but as np.nan.

This commit adds some logic to  savely handle both versions of "None" interpreted as string or
as np.nan.