Completes addition of reading .pq and .parquet files in parsing.
Earlier commit (first commit of this branch) added .pq as an option
to parsing, but it also requires an adjustment from gpd.read_file
to gpd.read_parquet.

Note that further checks of "type_map" parameters (see previous commit)
have revealed that the parameter is depreciated and should be removed.