Mitigate incorrect source_file values in Flanders, Belgium
The source_file column for Flanders, Belgium contained an incorrectly
parsed value (flanders-3035-rawflanders-3035-raw), resulting in the
source dataset mapping to fail.

Adding this duplicate string to the source dataset mapping mitigates the
issue and correctly replaces it with the string "belgium-flanders".

However, the underlying issue originates from the database setup or
earlier, and is already present in
2-database-nuts-level-v1-gov/belgium/BE256.gpkg.